requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.4
uvicorn[standard]==0.32.0

# ── LangChain / LLM ───────────────────────────────────────────────────────────
langchain==0.3.7
langchain-openai==0.2.8
langchain-core==0.3.18

# ── Data / Finance ────────────────────────────────────────────────────────────
yfinance==0.2.50
requests==2.32.3

# ── SEC EDGAR ─────────────────────────────────────────────────────────────────
sec-edgar-downloader==5.0.3

# ── Validation & config ───────────────────────────────────────────────────────
pydantic==2.9.2
python-dotenv==1.0.1

# ── Async HTTP (used by uvicorn / FastAPI internally) ─────────────────────────
httpx==0.27.2
